fixing android build for Expo SDK update and react native updates (there's a lot of changes). also adding a default github action to publush a build (but will fail until I get a key added)
