+++ riverdata/android/app/src/main/res/values-night/colors.xml
@@ -1,0 +1,1 @@
+<resources/>
